requirements.txt
django_haystack==3.3.0
fastapi==0.115.12
motor==3.7.0
pydantic==2.11.3
pymongo==4.12.0
uvicorn==0.34.0
haystack-ai>=2.0.0
qdrant-client>=1.6.4
qdrant-haystack==9.1.1
fastembed-haystack==1.4.1
sentence-transformers>=3.0.0